Refactor InputHandler logic
