+++ src/Colors/Rgb/Color.php
@@ -9,7 +9,7 @@ use Intervention\Image\Colors\Rgb\Channels\Blue;
 use Intervention\Image\Colors\Rgb\Channels\Green;
 use Intervention\Image\Colors\Rgb\Channels\Red;
 use Intervention\Image\Colors\Rgb\Channels\Alpha;
-use Intervention\Image\Drivers\AbstractInputHandler;
+use Intervention\Image\InputHandler;
 use Intervention\Image\Interfaces\ColorChannelInterface;
 use Intervention\Image\Interfaces\ColorInterface;
 use Intervention\Image\Interfaces\ColorspaceInterface;
@@ -53,14 +53,12 @@ class Color extends AbstractColor
      */
     public static function create(mixed $input): ColorInterface
     {
-        return (new class ([
+        return (new InputHandler([
             Decoders\HexColorDecoder::class,
             Decoders\StringColorDecoder::class,
             Decoders\TransparentColorDecoder::class,
             Decoders\HtmlColornameDecoder::class,
-        ]) extends AbstractInputHandler
-        {
-        })->handle($input);
+        ]))->handle($input);
     }
 
     /**
