+++ src/Decoders/ColorObjectDecoder.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Intervention\Image\Decoders;
+
+use Intervention\Image\Drivers\SpecializableDecoder;
+
+class ColorObjectDecoder extends SpecializableDecoder
+{
+}
