+++ src/Drivers/Imagick/Decoders/ColorObjectDecoder.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Intervention\Image\Drivers\Imagick\Decoders;
 
-use Intervention\Image\Drivers\AbstractDecoder;
+use Intervention\Image\Drivers\SpecializableDecoder;
 use Intervention\Image\Exceptions\DecoderException;
 use Intervention\Image\Interfaces\ColorInterface;
 use Intervention\Image\Interfaces\ImageInterface;
 
-class ColorObjectDecoder extends AbstractDecoder
+class ColorObjectDecoder extends SpecializableDecoder
 {
     public function decode(mixed $input): ImageInterface|ColorInterface
     {
