+++ src/Drivers/Imagick/Driver.php
@@ -13,7 +13,6 @@ use Intervention\Image\Exceptions\RuntimeException;
 use Intervention\Image\Format;
 use Intervention\Image\FileExtension;
 use Intervention\Image\Image;
-use Intervention\Image\Interfaces\ColorInterface;
 use Intervention\Image\Interfaces\ColorProcessorInterface;
 use Intervention\Image\Interfaces\ColorspaceInterface;
 use Intervention\Image\Interfaces\DriverInterface;
@@ -114,16 +113,6 @@ class Driver extends AbstractDriver
         $init($animation);
 
         return call_user_func($animation);
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @see DriverInterface::handleInput()
-     */
-    public function handleInput(mixed $input, array $decoders = []): ImageInterface|ColorInterface
-    {
-        return (new InputHandler($this->specializeMultiple($decoders)))->handle($input);
     }
 
     /**
