+++ tests/GdTestCase.php
@@ -14,7 +14,7 @@ abstract class GdTestCase extends BaseTestCase
 {
     public function readTestImage($filename = 'test.jpg'): Image
     {
-        return (new FilePathImageDecoder())->handle(
+        return (new FilePathImageDecoder())->decode(
             $this->getTestResourcePath($filename)
         );
     }
